Updated sigil parser

+++ D4DataParser/Entities/SigilInfo.cs
@@ -8,7 +8,7 @@ namespace D4DataParser.Entities
 {
     public class SigilInfo
     {
-        public int IdSno { get; set; }
+        public long IdSno { get; set; }
         public string IdName { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
